morphing all c to c++ to match link, reach compile but no link, need tons of new osd_ api + drivers

+++ mame-triumph/src/amiga/gui.h
@@ -15,6 +15,6 @@
 
 void AllocGUI(void);
 void FreeGUI(void);
-LONG MainGUI(void);
+int  MainGUI(void);
 void AboutGUI(void);
 #endif
